08_AIヒット予測システム/03_論文データ/beauty-health-trend-tracker/requirements-cloud.txt
# Core dependencies
requests==2.31.0
google-generativeai==0.3.2

# Web framework for Cloud Run
Flask==3.0.0
gunicorn==21.2.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Google Cloud (optional)
google-cloud-storage==2.10.0  # For GCS integration

# Environment management
python-dotenv==1.0.0
